all working good

+++ inc/client.hpp
@@ -60,7 +60,7 @@ struct Multipart_ENV
     std::string file_name;
     std::string content_type;
     std::string data;
-    int pos;
+    unsigned long pos;
 };
